// src/components/Uploader/enum.ts
export enum UploadWays {
  Image = 'image',
  Camera = 'camera',
  File = 'file',
}

export enum PhoneOS {
  Android = 'android',
  Ios = 'ios',
}

export enum PhoneAuths {
  Camera = 'Camera',
  PhotoLibrary = 'PhotoLibrary',
  Biometrics = 'Biometrics',
}

export enum CacheKeys {
  AndroidToPhoto = 'android_to_photo', // 安卓去拿照片
}

export enum UploadFileTypes {
  AllFile = 'allFiles',
  Image = 'images',
  Pdf = 'pdf',
  Audio = 'audio',
  Csv = 'csv',
  PlainText = 'plainText',
  Video = 'video',
  Zip = 'zip',
}

export enum UploadListTypes {
  Text = 'text', // 图片列表
  PictureCard = 'picture-card', // 图片卡片列表
}